+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `EtaLikeSimplifications`: Confusing help string for Boolean literals: Issue [#392](https://github.com/nowarp/misti/issues/392)
 - No output with `-o json` without `--output-path` specified: Issue [#403](https://github.com/nowarp/misti/issues/403)
 - `-o json --output-path <some dir>` stores warnings in the plain text format
+- `PreferGlobalFunction`: Don't report methods with attributes: Issue [#406](https://github.com/nowarp/misti/issues/406)
 
 ## [0.8.3] - 2025-04-26